#refactor: moved cors env and marker to separate package folder
