+++ src/core/error/index.ts
@@ -1,7 +1,7 @@
 import joorData from '@/data/joor';
 import JoorError from '@/core/error/JoorError';
 import { JOOR_ERROR } from '@/types/error';
-import marker from '@/core/marker';
+import marker from '@/packages/marker';
 
 /**
  * Class to work with errors
