+++ src/packages/cors/index.ts
